Monitoring/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
motor==3.6.0
python-socketio==5.11.4
aio-pika==9.5.4
python-dotenv==1.0.1
